Rename PocketLoggedOutViewModel to LoggedOutViewModel
This fits better with the consistently named "LoggedOutX" file and type naming.
